[refactor](schema change) refact fe light schema change.
